+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Notable changes to the BeeGFS CSI driver will be documented in this file.
 ### Added
 - Support for Kubernetes 1.26 and 1.27.
 - Support for BeeGFS v7.3.4 and v7.4.0.
+- Support for binary connAuthFile secrets utilizing base64 encoding.
 
 ### Changed
 - Migrated project to the ThinkParQ GitHub organization.